requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
ultralytics>=8.0.0
opencv-python>=4.8.0
google-search-results>=2.4.2
python-dotenv>=1.0.0
cachetools>=5.3.2
numpy>=1.24.0
Pillow>=10.0.0
langchain>=0.0.350
langchain-community>=0.0.10
crewai>=0.10.0
duckduckgo-search>=4.1.1
pydantic>=2.0.0
litellm>=1.16.9
serpapi>=2.15.0
